Address rectilinear vector calculus review
