+++ physicsnemo/nn/functional/derivatives/rectilinear_grid_laplacian/_warp_impl/_kernels/__init__.py
@@ -1,0 +1,35 @@
+# SPDX-FileCopyrightText: Copyright (c) 2023 - 2026 NVIDIA CORPORATION & AFFILIATES.
+# SPDX-FileCopyrightText: All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .backward import (
+    _laplacian_backward_1d_kernel,
+    _laplacian_backward_2d_kernel,
+    _laplacian_backward_3d_kernel,
+)
+from .forward import (
+    _laplacian_1d_kernel,
+    _laplacian_2d_kernel,
+    _laplacian_3d_kernel,
+)
+
+__all__ = [
+    "_laplacian_1d_kernel",
+    "_laplacian_2d_kernel",
+    "_laplacian_3d_kernel",
+    "_laplacian_backward_1d_kernel",
+    "_laplacian_backward_2d_kernel",
+    "_laplacian_backward_3d_kernel",
+]
